fix(YALB-520): use link-style attribute
It looks as if the link line bleeds into the spacing if the link-style
is not set as a data attribute.  Seeing as the storybook version has
this always underlined, I'm defaulting to this.

The only issue is that when doing this, the link breaks to its own line.
I believe this to be more of a component-library-twig issue offhand, but
will look into it more as part of this.