fix(YALB-520): fix newlines for links
When using the underline style, it causes the link to be on its own
line.  While is is good for certain cases, for inline text from the
user, this can be a bad thing.  This is an attempt to address this.
It feels a bit dirty, but wanted to get this down for now.