fix(YALB-520): copy stripped mailto link
The component is copying from the span, but in our case the email
address might not always be the text on the screen.  This creates a new
hidden span inside the link to store the real email address so that it
can copy that instead when the copy button is clicked.