fix(YALB-520): add weight and name to link types
Feedback was given that in certain circumstances, the links were out of
order, resulting in false treatments of certain types.  A weight has
been implemented that will force the order of the evaluations such that
the correct one should be selected.  The name was also added as an
easier way of determining the link type being evaluated and what weight
order resulted when debugging is turned on.

This also introduces treatments to anchors (the chevron) and the blank
target.  Even thought the current Drupal setup does not support blank
targets, this is an upcoming feature, and so it has been enabled.