fix(YALB-520): prep for component library update
In the storybook examples of component library twig, we do not currently
use the hidden span to have the mailto link.  This allows us to use
either to accomplish this.

If the hidden span exists, we use it, otherwise we assume that the span
before it should be used, giving us the backward compatible
functionality that will work in component library twig.