fix(YALB-520): move copy into span
This is an attempt to clean up instances where the copy is inside of a
component that has no wrapper around it.  This sets up the possibility
that if the link is inside of a text paragraph, it can still display the
old styling.

This might change in the future into something else, like a pill, but
it's a start.